Hasher and similarity logic and benchmark metrics (#9)
Initial features of material-hasher package, which includes logic for passing various benchmarks (lattice transformations, and disordered materials) on various "similarity" and "hasher" methods.

**Introduction of two base classes:**
Similarity methods are defined to be a method which does not fingerprint or hash a inorganic crystalline structure (ie returns a string) but rather can distinguish whether two structures are equivalent or not to a given threshold.

Both `similarity` and `hasher` classes have the following methods:
- `is_equivalent` (for hash based classes this is determined to be if the fingerprint of each structure are equivalent, while for similarity, it is if they are close enough to a default threshold)
- `get_pairwise_equivalence` (same as above but for a list of structures)

A hash class must have a `get_material_hash` method implemented which takes a Pymatgen Structure object and outputs a fingerprint for it. Meanwhile a similarity class but have a `get_similarity_score` method implemented which can take two structures and determine a score between them.

**In this release, we include the following fingerprinting methods:**
- CLOUD
- SLICES *requires separate package install*
- Entalpic
- PDD

**And we include two similarity method:**
- Pymatgen
- EquiformerV2

In order to benchmark these methods, we release a set of benchmark under the `benchmark` module. Their are two main benchmarks: disordered and transformations. Transformations are meant to capture both the level of noise (atomic coordinates, lattice vectors) prior to a new hash from a hasher class or prior to detecting a novel structure from a similarity class; as well as test for a closed set of affine transformation to the lattice (symmetry operations within the symmetry detected in the structure and lattice translations).
**Includes the following benchmarks:**
- disordered structures
- transformations
-- noise on atomic position
-- noise on lattice vectors
-- strain on lattice
-- symmetry operations of lattice
-- translations

In order to test these methods, on these set of benchmark we make use of LeMat-Bulk dataset in Huggingface and bring appropriate methods to access the data, sample it, and load it as a list of Pymatgen Structure objects.